don't allocate 1mb on the stack, rename .h
